note title update moved to editor, but bad UX
